Title: 가장 많이 겹치는 구간 2, Time: -1ms, Memory: -1MB, Status: Compile Error - Codetree

+++ 260314/가장 많이 겹치는 구간 2/section-with-maximum-overlap-2.cpp
@@ -7,8 +7,8 @@ long long x1[100004], x2[100004];
 long long arr[100004];
 //1000 000 000
 int mx = INT_MIN;
-int ans = INT_MIN;
-int sum;
+long long ans = INT_MIN;
+long sum;
 int main() {
     cin >> n;
     for (int i = 0; i < n; i++) {
